% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{add_all_dependencies}
\alias{add_all_dependencies}
\title{Attach all created dependencies in the ./R directory to the tag}
\usage{
add_all_dependencies(tag)
}
\arguments{
\item{tag}{Tag to attach the dependencies}
}
\description{
Attach all created dependencies in the ./R directory to the tag
}
\examples{
\dontrun{
 library(htmltools)
 findDependencies(add_all_dependencies(div()))
}
}
