% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{add_dependencies}
\alias{add_dependencies}
\title{Attach all created dependencies in the ./R directory to the tag}
\usage{
add_dependencies(tag, deps = NULL)
}
\arguments{
\item{tag}{Tag to attach the dependencies}

\item{deps}{Dependencies to add. Expect a vector of names. If NULL, all dependencies
are added.}
}
\description{
This function only works if there are existing dependencies. Otherwise,
an error is raised.
}
\examples{
\dontrun{
 library(htmltools)
 findDependencies(add_dependencies(div()))
 findDependencies(add_dependencies(div(), deps = "bulma"))
}
}
