% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{charpente_options}
\alias{charpente_options}
\title{Configure charpente}
\usage{
charpente_options(local = TRUE)
}
\arguments{
\item{local}{Whether to download files locally or to point to a CDN. Default to TRUE.}
}
\value{
A list of options.
}
\description{
Configure charpente
}
