% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_files.R
\name{compress_js}
\alias{compress_js}
\title{Compress and optimize all files in the current folder}
\usage{
compress_js(dir = "srcjs", source_maps = TRUE)
}
\arguments{
\item{dir}{Default to srcjs.}

\item{source_maps}{Default to TRUE.}
}
\description{
Generates a minified file under inst/pkg_name-pkg_version.
}
