% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_charpente_utils}
\alias{copy_charpente_utils}
\title{Copy charpente utils function}
\usage{
copy_charpente_utils(pkg_name)
}
\value{
Copy all charpente utils function in the ./R/PKG_NAM-utils.R script
}
\description{
Copy charpente utils function
}
\keyword{internal}
