% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charpente.R
\name{create_charpente}
\alias{create_charpente}
\title{Create a package using charpente and usethis conventions}
\usage{
create_charpente(path, remote = NULL, private = FALSE, license)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is created,
provided that the parent path exists.}

\item{remote}{Name of the remote Github organization to connect to.}

\item{private}{Whether the repository is private. Default to FALSE.}

\item{license}{Which license is your project under?}
}
\description{
A charpente powered package does not differ from other package structure.
\link{create_charpente} calls a series of function to help you quickly setting up the
package structure.
}
\examples{
\dontrun{
 create_charpente("bs4Dash")
}
}
