% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{create_custom_dependency}
\alias{create_custom_dependency}
\title{Imports Internal Dependencies}
\usage{
create_custom_dependency(
  name,
  script = NULL,
  stylesheet = NULL,
  open = interactive()
)
}
\arguments{
\item{name}{Name of library.}

\item{script}{List of scripts to include.}

\item{stylesheet}{List of stylesheets to include.}

\item{open}{Whether to allow rstudioapi to open the newly created script. Default to TRUE.}
}
\description{
Wrap internal scripts and stylesheets in one \link[htmltools]{htmlDependency}.
}
\examples{
\dontrun{
 create_custom_dependency("custom", script = c("hello.js", "blabla.js"))
}
}
