% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{create_dependency}
\alias{create_dependency}
\title{Imports External Dependencies}
\usage{
create_dependency(name, local = TRUE, version = NULL)
}
\arguments{
\item{name}{Name of library.}

\item{local}{Whether to download files locally or to point to a CDN. Default to TRUE.}

\item{version}{Library version. Default to NULL. If NULL, takes the latest version.}
}
\description{
Download and create dependency function.
}
