% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwa.R
\name{create_manifest}
\alias{create_manifest}
\title{Create a manifest for your shiny app}
\usage{
create_manifest(
  path,
  name = "My Progressive Web App",
  shortName = "My App",
  description = "What it does!",
  lang = "en-US",
  startUrl = "/",
  display = c("standalone", "minimal-ui", "fullscreen", "browser"),
  background_color = "#ffffff",
  theme_color = "#ffffff"
)
}
\arguments{
\item{path}{App path.}

\item{name}{App name.}

\item{shortName}{App short name.}

\item{description}{App description}

\item{lang}{App language (en-US by default).}

\item{startUrl}{Page to open at start.}

\item{display}{Display mode. Choose among \code{c("minimal-ui", "standalone", "fullscreen", "browser")}.
In practice, you want the standalone mode so that the app looks like a native app.}

\item{background_color}{The background_color property is used on the splash screen when the application is first launched.}

\item{theme_color}{The theme_color sets the color of the tool bar, and may be reflected in the app's preview in task switchers.}
}
\value{
This function creates a www folder for your shiny app. Must specify the path.
It creates 1 folders to contain icons and the manifest.json file.
}
\description{
This is a central piece if you want to have your app standalone for instance
}
\note{
See \url{https://developer.mozilla.org/en-US/docs/Web/Manifest} for more informations.
}
\examples{
create_manifest(
  path = tempdir(),
  name = "My App",
  shortName = "My App",
  description = "What it does!",
  lang = "en-US",
  startUrl = "https://www.google.com/",
  display = "standalone",
  background_color = "#3367D6",
  theme_color = "#3367D6"
)
}
