% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwa.R
\name{create_pwa_dependency}
\alias{create_pwa_dependency}
\title{Create a PWA dependency}
\usage{
create_pwa_dependency(open = interactive())
}
\arguments{
\item{open}{Whether to allow rstudioapi to open the newly created script. Default to TRUE.}
}
\description{
List all relevant pwa resources in a custom dependency. Includes
the web manifest + icons using the head parameter of htmltools::htmlDependency.
}
