% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{get_dependency_assets}
\alias{get_dependency_assets}
\title{Get all links to dependencies}
\usage{
get_dependency_assets(dep, tag = "latest")
}
\arguments{
\item{dep}{Library name.}

\item{tag}{Library version. Default to latest.}
}
\value{
A list of url containing links to CSS and JS dependencies for the given library.
}
\description{
Query from \url{https://data.jsdelivr.com/v1/package/npm/} under the hood.
}
\examples{
\dontrun{
 get_dependency_assets("bootstrap")
 get_dependency_assets("framework7", tag = "5.5.5")
}
}
