% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{get_installed_dependency}
\alias{get_installed_dependency}
\title{Get the version of the current installed dependency}
\usage{
get_installed_dependency(name)
}
\arguments{
\item{name}{Library name}
}
\value{
A character containing the version number
}
\description{
Used by \link{update_dependency}.
}
