% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_template}
\alias{get_template}
\title{Return the full path for a default file template}
\usage{
get_template(file, where)
}
\arguments{
\item{file}{Name of the file including the file extension.}

\item{where}{Where to find the template.}
}
\description{
Return the full path for a default file template
}
\keyword{internal}
