% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_template}
\alias{process_template}
\title{Insert provided parameters into template}
\usage{
process_template(
  template,
  ...,
  where = system.file("utils", package = "charpente")
)
}
\arguments{
\item{template}{Template obtained with \link{get_template}.}

\item{...}{List of parameters to insert in the template.}

\item{where}{Where to find the template.}
}
\description{
Insert provided parameters into template
}
\keyword{internal}
