% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reference_script}
\alias{reference_script}
\title{Add script.js to main.js entry point.}
\usage{
reference_script(name)
}
\arguments{
\item{name}{Script name.}
}
\description{
Needed to keep the entry point up to date
}
\keyword{internal}
