% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reference_style}
\alias{reference_style}
\title{Insert import into main SCSS file}
\usage{
reference_style(path)
}
\arguments{
\item{path}{Could be file.name or folder/file.name}
}
\description{
Useful to reference scss files into the main one.
}
\keyword{internal}
