% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_esbuild}
\alias{set_esbuild}
\title{Setup esbuild}
\usage{
set_esbuild(light = FALSE)
}
\arguments{
\item{light}{Used to only install Sass plugins. This is
to workaround a breaking change in charpente where styles does
not exist in old versions.}
}
\value{
Installs esbuild in node_modules (dev scope), creates srcjs + srcjs/main.js
}
\description{
Installs esbuild for the local project
}
\keyword{internal}
