% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_mocha}
\alias{set_mocha}
\title{Setup mocha}
\usage{
set_mocha()
}
\value{
Installs mocha in node_modules (dev scope), creates srcjs/test folder,
write basic test im test_basic.js
}
\description{
Installs mocha for the local project
}
\keyword{internal}
