% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwa.R
\name{set_pwa}
\alias{set_pwa}
\title{Utils to set up a PWA compatible structure}
\usage{
set_pwa(
  path,
  name = "My Progressive Web App",
  shortName = "My App",
  description = "What it does!",
  lang = "en-US",
  startUrl = "/",
  display = c("standalone", "minimal-ui", "fullscreen", "browser"),
  background_color = "#ffffff",
  theme_color = "#ffffff",
  register_service_worker = TRUE,
  create_dependencies = TRUE
)
}
\arguments{
\item{path}{App path.}

\item{name}{App name.}

\item{shortName}{App short name.}

\item{description}{App description}

\item{lang}{App language (en-US by default).}

\item{startUrl}{Page to open at start.}

\item{display}{Display mode. Choose among \code{c("minimal-ui", "standalone", "fullscreen", "browser")}.
In practice, you want the standalone mode so that the app looks like a native app.}

\item{background_color}{The background_color property is used on the splash screen when the application is first launched.}

\item{theme_color}{The theme_color sets the color of the tool bar, and may be reflected in the app's preview in task switchers.}

\item{register_service_worker}{Whether to register the service worker. Default to
TRUE. Don't change the file name of service-worker.js!!!}

\item{create_dependencies}{Default to TRUE. Relevant if used in a shinyMobile context.
If used outside, you must set it to FALSE and handle the dependencies
yourself.}
}
\description{
Creates a web manifest, service-worker.js, icons and set the
necessary dependencies. The app must be part of a package.
Must not be used from the package root but from the app root.
}
