% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_version_control}
\alias{set_version_control}
\title{Setup version control}
\usage{
set_version_control(remote, private)
}
\arguments{
\item{remote}{Name of the remote Github organization to connect to.}

\item{private}{Whether the repository is private. Default to FALSE.}
}
\description{
Sets git and optionally link to github for the current project
}
\keyword{internal}
