% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{update_dependency}
\alias{update_dependency}
\title{Update the given dependency to a specific version or latest}
\usage{
update_dependency(name, version_target = "latest")
}
\arguments{
\item{name}{Library name.}

\item{version_target}{Targeted version. Default to latest.}
}
\description{
Update the given dependency to a specific version or latest
}
\examples{
\dontrun{
 update_dependency("framework7")
}
}
