% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{upgrade_dependency}
\alias{upgrade_dependency}
\title{Upgrade the given dependency to a specific version or latest}
\usage{
upgrade_dependency(name, version_target = "latest")
}
\arguments{
\item{name}{Library name.}

\item{version_target}{Targeted version. Defaul to latest.}
}
\description{
Upgrade the given dependency to a specific version or latest
}
\examples{
\donttun{
 upgrade_dependency("framework7")
}
}
