% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_files.R
\name{build_js}
\alias{build_js}
\title{Compress and optimize all files in the current folder}
\usage{
build_js(
  dir = "srcjs",
  files = NULL,
  mode = c("prod", "dev"),
  source_maps = TRUE,
  ...
)
}
\arguments{
\item{dir}{Default to srcjs.}

\item{files}{Use to specify custom order for files processing.
Default to NULL.}

\item{mode}{Production or development mode. Choose either "prod" or "dev".}

\item{source_maps}{Default to TRUE. Whether to include sourceMaps
for minified scripts. Default to TRUE.}

\item{...}{Other options. See \url{https://github.com/terser/terser}.}
}
\description{
Generates a minified file under inst/pkg_name-pkg_version, if mode
is prod. If mode is dev, aggregates all js files without mangling
or compression.
}
