% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{create_custom_dependency}
\alias{create_custom_dependency}
\title{Imports Internal Dependencies}
\usage{
create_custom_dependency(
  name,
  version,
  script = NULL,
  stylesheet = NULL,
  open = interactive(),
  mode
)
}
\arguments{
\item{name}{Package name.}

\item{version}{Package version.}

\item{script}{List of scripts to include. Assumes scripts are located at the root of the inst folder.}

\item{stylesheet}{List of stylesheets to include. Assumes stylesheets are located at the root of the inst folder.}

\item{open}{Whether to allow rstudioapi to open the newly created script. Default to TRUE.}

\item{mode}{Internal. Don't use.}
}
\description{
Wrap internal scripts and stylesheets in one \link[htmltools]{htmlDependency}.
}
\examples{
\dontrun{
 create_custom_dependency("custom", script = c("hello.js", "blabla.js"))
}
}
\keyword{Internal}
