% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{create_custom_dependency}
\alias{create_custom_dependency}
\title{Imports Internal Dependencies}
\usage{
create_custom_dependency(
  name,
  version,
  entry_points,
  open = interactive(),
  mode
)
}
\arguments{
\item{name}{Package name.}

\item{version}{Package version.}

\item{entry_points}{Entry points to create dependency for.}

\item{open}{Whether to allow rstudioapi to open the newly created script. Default to TRUE.}

\item{mode}{Internal. Don't use.}
}
\description{
Wrap internal scripts and stylesheets in one \link[htmltools]{htmlDependency}.
}
\keyword{Internal}
