% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{create_dependency}
\alias{create_dependency}
\title{Imports External Dependencies}
\usage{
create_dependency(name, version = NULL, options = charpente_options())
}
\arguments{
\item{name}{Name of library.}

\item{version}{Library version. Default to NULL. If NULL, takes the latest version.}

\item{options}{See \link{charpente_options}.}
}
\description{
Download and create dependency function.
}
