% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{create_dependency}
\alias{create_dependency}
\title{Imports External Dependencies}
\usage{
create_dependency(
  name,
  pkg,
  tag = NULL,
  open = TRUE,
  options = charpente_options()
)
}
\arguments{
\item{name}{Name of library.}

\item{pkg}{Package where to create the dependency.}

\item{tag}{Library version. Default to NULL. If NULL, takes the latest version.}

\item{open}{Whether to allow rstudioapi to open the newly created script. Default to TRUE.}

\item{options}{See \link{charpente_options}.}
}
\description{
Download and create dependency function.
}
\examples{
\dontrun{
 create_dependency("tabler", pkg = "tablerDash")
 # Use CDNs
 create_dependency(
  "framework7",
  pkg = "shinyMobile",
  options = charpente_options(local = FALSE)
 )
}
}
