% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{create_dependency}
\alias{create_dependency}
\title{Imports External Dependencies}
\usage{
create_dependency(name, scripts = NULL, stylesheets = NULL)
}
\arguments{
\item{name}{Name of library.}

\item{scripts, stylesheets}{Vector of JavaScript and CSS files to use
as depdendencies.}
}
\description{
Download and create depdendency function.
}
