% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_files.R
\name{create_input_binding}
\alias{create_input_binding}
\alias{create_output_binding}
\alias{create_custom_handler}
\alias{create_js}
\alias{create_css}
\title{Create a shiny custom input binding boilerplate}
\usage{
create_input_binding(...)

create_output_binding(...)

create_custom_handler(...)

create_js(...)

create_css(...)
}
\arguments{
\item{...}{Arguments to be passed to the \code{module_template} function.}
}
\description{
Create a shiny custom input binding boilerplate

Create a shiny output binding boilerplate

Create a shiny custom handler boilerplate

Create a JavaScript file

Create a css file
}
