% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{get_dependency_versions}
\alias{get_dependency_versions}
\title{Get all version for the current dependency}
\usage{
get_dependency_versions(dep, cdn = NULL, latest = FALSE)
}
\arguments{
\item{dep}{Library name.}

\item{cdn}{CDN url. Default to \url{https://api.cdnjs.com}.}

\item{latest}{Whether to get the last version. Default to FALSE.}
}
\value{
A vector containing all versions
}
\description{
Get all version for the current dependency
}
\examples{
get_dependency_versions("framework7")
get_dependency_versions("bootstrap")
get_dependency_versions("react", latest = TRUE)
}
