% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_esbuild}
\alias{run_esbuild}
\title{Run esbuild}
\usage{
run_esbuild(mode, outputDir)
}
\arguments{
\item{mode}{Production or development mode. Choose either "prod" or "dev".
"prod" bundles, aggregates and minifyies files. "dev" only bundles the code.
Modules follow the ES6 format (import/export).}

\item{outputDir}{Output directory}
}
\description{
Apply esbuild to the srcjs folder.
}
\keyword{internal}
